feat: Separate integration tests into staging and release
