+++ .github/workflows/test-release.yml
@@ -1,6 +1,9 @@
 name: test
 
-on: push
+on:
+  push:
+    branches:
+      - master
 
 jobs:
     build:
@@ -14,4 +17,4 @@ jobs:
             - name: Run and set up the parachain, oracle, staked relayer and vault
               run: docker-compose up -d
             - run: yarn install
-            - run: yarn ci:test
+            - run: yarn ci:test:release
