+++ .github/workflows/test-staging.yml
@@ -1,0 +1,20 @@
+name: test
+
+on:
+  pull_request:
+    branches:
+      - master
+
+jobs:
+    build:
+        runs-on: ubuntu-latest
+        steps:
+            - uses: actions/checkout@v2
+            - name: setup node
+              uses: actions/setup-node@v1
+              with:
+                  node-version: "14.x"
+            - name: Run and set up the parachain, oracle, staked relayer and vault
+              run: docker-compose up -d
+            - run: yarn install
+            - run: yarn ci:test:staging
