+++ package.json
@@ -17,7 +17,8 @@
     "fix": "run-s fix:*",
     "fix:prettier": "prettier \"src/**/*.ts\" --write",
     "fix:lint": "eslint --fix . --ext .ts",
-    "ci:test": "run-s build test:lint test:unit test:integration",
+    "ci:test:staging": "run-s build test:lint test:unit test:integration:staging",
+    "ci:test:release": "run-s build test:integration:release",
     "ci:test-with-coverage": "nyc -r lcov -e .ts -x \"*.test.ts\" yarn ci:test",
     "docs": "./generate_docs",
     "generate:defs": "ts-node --skip-project node_modules/.bin/polkadot-types-from-defs --package @interlay/polkabtc/interfaces --input ./src/interfaces",
@@ -25,7 +26,8 @@
     "test": "run-s build test:*",
     "test:lint": "eslint src --ext .ts",
     "test:unit": "mocha test/unit/*.test.ts test/unit/**/*.test.ts",
-    "test:integration": "mocha test/integration/**/*.test.ts --timeout 10000000",
+    "test:integration:staging": "mocha test/integration/**/staging/*.test.ts --timeout 10000000",
+    "test:integration:release": "mocha test/integration/**/release/*.test.ts --timeout 10000000",
     "watch:build": "tsc -p tsconfig.json -w",
     "watch:test": "mocha --watch test/**/*.test.ts"
   },
