+++ test/integration/parachain/staging/btc-relay.test.ts
@@ -1,10 +1,10 @@
 import { ApiPromise } from "@polkadot/api";
 import { assert } from "chai";
-import { BTCRelayAPI } from "../../../src/parachain";
-import { BTCCoreAPI, DefaultBTCCoreAPI } from "../../../src/external/btc-core";
-import { DefaultBTCRelayAPI } from "../../../src/parachain/btc-relay";
-import { createPolkadotAPI } from "../../../src/factory";
-import { defaultParachainEndpoint } from "../../config";
+import { BTCRelayAPI } from "../../../../src/parachain";
+import { BTCCoreAPI, DefaultBTCCoreAPI } from "../../../../src/external/btc-core";
+import { DefaultBTCRelayAPI } from "../../../../src/parachain/btc-relay";
+import { createPolkadotAPI } from "../../../../src/factory";
+import { defaultParachainEndpoint } from "../../../config";
 
 describe("BTCCore", function () {
     this.timeout(10000); // API can be slightly slow
