+++ test/integration/parachain/staging/oracle.test.ts
@@ -1,9 +1,9 @@
 import { ApiPromise, Keyring } from "@polkadot/api";
 import { KeyringPair } from "@polkadot/keyring/types";
-import { DefaultOracleAPI, OracleAPI } from "../../../src/parachain/oracle";
-import { createPolkadotAPI } from "../../../src/factory";
-import { assert } from "../../chai";
-import { defaultParachainEndpoint } from "../../config";
+import { DefaultOracleAPI, OracleAPI } from "../../../../src/parachain/oracle";
+import { createPolkadotAPI } from "../../../../src/factory";
+import { assert } from "../../../chai";
+import { defaultParachainEndpoint } from "../../../config";
 
 describe("OracleAPI", () => {
     let api: ApiPromise;
