+++ test/integration/parachain/staging/staked-relayer.test.ts
@@ -2,11 +2,11 @@ import { ApiPromise, Keyring } from "@polkadot/api";
 import { AccountId } from "@polkadot/types/interfaces/runtime";
 import BN from "bn.js";
 import sinon from "sinon";
-import { DefaultStakedRelayerAPI, StakedRelayerAPI } from "../../../src/parachain/staked-relayer";
-import { createPolkadotAPI } from "../../../src/factory";
-import { StakedRelayer, DOT } from "../../../src/interfaces/default";
-import { assert } from "../../chai";
-import { defaultParachainEndpoint } from "../../config";
+import { DefaultStakedRelayerAPI, StakedRelayerAPI } from "../../../../src/parachain/staked-relayer";
+import { createPolkadotAPI } from "../../../../src/factory";
+import { StakedRelayer, DOT } from "../../../../src/interfaces/default";
+import { assert } from "../../../chai";
+import { defaultParachainEndpoint } from "../../../config";
 import * as bitcoin from "bitcoinjs-lib";
 import { KeyringPair } from "@polkadot/keyring/types";
 import Big from "big.js";
